workspace/config/dev/requirements.txt
# 企业级开发环境Python依赖
requests>=2.28.0
pyyaml>=6.0
kubernetes>=24.0
docker>=6.0
jsonschema>=4.0
cryptography>=38.0
python-dotenv>=0.19.0
pytest>=7.0
pytest-asyncio>=0.18.0
black>=22.0
flake8>=5.0
mypy>=0.982